+++ disma-cli/src/infra/config/role.rs
@@ -115,122 +115,136 @@ impl Into<AwaitingRole> for RoleConfig {
 
 #[cfg(test)]
 mod test {
-    use disma::{
-        permission::{Permission, PermissionsList},
-        role::{AwaitingRole, ExistingRole},
-    };
 
-    use super::RoleConfig;
-
-    #[test]
-    fn can_convert_config_to_awaiting_entity() {
-        let is_mentionable = true;
-        let show_in_sidebar = false;
-        let color = "826d5f".to_string();
-        let name = "Team10".to_string();
-
-        let config = RoleConfig {
-            name: name.clone(),
-            color: Some(color.clone()),
-            show_in_sidebar,
-            is_mentionable,
-            permissions: Some(vec!["ADMINISTRATOR".to_string()]),
+    mod into_awaiting {
+        use disma::{
+            permission::{Permission, PermissionsList},
+            role::AwaitingRole,
         };
 
-        let entity: AwaitingRole = config.into();
+        use crate::infra::config::role::RoleConfig;
 
-        let expected_entity = AwaitingRole {
-            name: name.clone(),
-            color: Some(color.clone()),
-            is_mentionable,
-            show_in_sidebar,
-            permissions: PermissionsList::from(vec![Permission::ADMINISTRATOR]),
-        };
-        assert_eq!(entity, expected_entity);
+        #[test]
+        fn can_convert_config_to_awaiting_entity() {
+            let is_mentionable = true;
+            let show_in_sidebar = false;
+            let color = "826d5f".to_string();
+            let name = "Team10".to_string();
+
+            let config = RoleConfig {
+                name: name.clone(),
+                color: Some(color.clone()),
+                show_in_sidebar,
+                is_mentionable,
+                permissions: Some(vec!["ADMINISTRATOR".to_string()]),
+            };
+
+            let entity: AwaitingRole = config.into();
+
+            let expected_entity = AwaitingRole {
+                name: name.clone(),
+                color: Some(color.clone()),
+                is_mentionable,
+                show_in_sidebar,
+                permissions: PermissionsList::from(vec![Permission::ADMINISTRATOR]),
+            };
+            assert_eq!(entity, expected_entity);
+        }
+
+        #[test]
+        fn can_convert_compressed_config_to_awaiting_entity() {
+            let is_mentionable = true;
+            let show_in_sidebar = false;
+            let name = "Team10".to_string();
+            let permissions: Vec<String> = vec![];
+
+            let config = RoleConfig {
+                name: name.clone(),
+                color: None,
+                is_mentionable,
+                show_in_sidebar,
+                permissions: None,
+            };
+
+            let entity: AwaitingRole = config.into();
+
+            let expected_entity = AwaitingRole {
+                name: name.clone(),
+                color: None,
+                is_mentionable,
+                show_in_sidebar,
+                permissions: PermissionsList::from(permissions),
+            };
+            assert_eq!(entity, expected_entity);
+        }
     }
 
-    #[test]
-    fn can_convert_compressed_config_to_awaiting_entity() {
-        let is_mentionable = true;
-        let show_in_sidebar = false;
-        let name = "Team10".to_string();
-        let permissions: Vec<String> = vec![];
-
-        let config = RoleConfig {
-            name: name.clone(),
-            color: None,
-            is_mentionable,
-            show_in_sidebar,
-            permissions: None,
+    mod from_existing {
+        use disma::{
+            permission::{Permission, PermissionsList},
+            role::ExistingRole,
         };
 
-        let entity: AwaitingRole = config.into();
+        use crate::infra::config::role::RoleConfig;
 
-        let expected_entity = AwaitingRole {
-            name: name.clone(),
-            color: None,
-            is_mentionable,
-            show_in_sidebar,
-            permissions: PermissionsList::from(permissions),
-        };
-        assert_eq!(entity, expected_entity);
+        #[test]
+        fn can_convert_existing_entity_to_config() {
+            let is_mentionable = true;
+            let show_in_sidebar = false;
+            let color = "826d5f".to_string();
+            let name = "Team10".to_string();
+            let id = "93jdi0".to_string();
+
+            let entity = ExistingRole {
+                id: id.clone(),
+                name: name.clone(),
+                color: Some(color.clone()),
+                is_mentionable,
+                show_in_sidebar,
+                permissions: PermissionsList::from(vec![Permission::ADMINISTRATOR]),
+            };
+
+            let config = RoleConfig::from(&entity);
+
+            let expected_config = RoleConfig {
+                name: name.clone(),
+                color: Some(color.clone()),
+                show_in_sidebar,
+                is_mentionable,
+                permissions: Some(vec!["ADMINISTRATOR".to_string()]),
+            };
+            assert_eq!(config, expected_config);
+        }
+
+        #[test]
+        fn can_convert_existing_entity_to_compressed_config() {
+            let is_mentionable = true;
+            let show_in_sidebar = false;
+            let name = "Team10".to_string();
+            let id = "93jdi0".to_string();
+            let permissions: Vec<String> = vec![];
+
+            let entity = ExistingRole {
+                id: id.clone(),
+                name: name.clone(),
+                color: None,
+                is_mentionable,
+                show_in_sidebar,
+                permissions: PermissionsList::from(permissions),
+            };
+
+            let config = RoleConfig::from(&entity);
+
+            let expected_config = RoleConfig {
+                name: name.clone(),
+                color: None,
+                show_in_sidebar,
+                is_mentionable,
+                permissions: None,
+            };
+            assert_eq!(config, expected_config);
+        }
     }
 
-    #[test]
-    fn can_convert_existing_entity_to_config() {
-        let is_mentionable = true;
-        let show_in_sidebar = false;
-        let color = "826d5f".to_string();
-        let name = "Team10".to_string();
-        let id = "93jdi0".to_string();
-
-        let entity = ExistingRole {
-            id: id.clone(),
-            name: name.clone(),
-            color: Some(color.clone()),
-            is_mentionable,
-            show_in_sidebar,
-            permissions: PermissionsList::from(vec![Permission::ADMINISTRATOR]),
-        };
-
-        let config = RoleConfig::from(&entity);
-
-        let expected_config = RoleConfig {
-            name: name.clone(),
-            color: Some(color.clone()),
-            show_in_sidebar,
-            is_mentionable,
-            permissions: Some(vec!["ADMINISTRATOR".to_string()]),
-        };
-        assert_eq!(config, expected_config);
-    }
-
-    #[test]
-    fn can_convert_existing_entity_to_compressed_config() {
-        let is_mentionable = true;
-        let show_in_sidebar = false;
-        let name = "Team10".to_string();
-        let id = "93jdi0".to_string();
-        let permissions: Vec<String> = vec![];
-
-        let entity = ExistingRole {
-            id: id.clone(),
-            name: name.clone(),
-            color: None,
-            is_mentionable,
-            show_in_sidebar,
-            permissions: PermissionsList::from(permissions),
-        };
-
-        let config = RoleConfig::from(&entity);
-
-        let expected_config = RoleConfig {
-            name: name.clone(),
-            color: None,
-            show_in_sidebar,
-            is_mentionable,
-            permissions: None,
-        };
-        assert_eq!(config, expected_config);
-    }
+    mod serde_parsing {}
 }
