+++ disma/Cargo.toml
@@ -12,9 +12,11 @@ keywords = ["discord", "cli", "config"]
 categories = ["command-line-utilities", "config"]
 
 [dependencies]
-mock-it = "0.9.0"
 reqwest = { version = "0.11.11", features = ["json", "blocking"] }
 serde = { version = "1.0.143", features = ["derive"] }
 serde_json = "1.0.83"
 serde_repr = "0.1.9"
 strum = { version = "0.24.1", features = ["derive"] }
+
+[dev-dependencies]
+mock-it = "0.9.0"
